+++ app.js
@@ -5,6 +5,7 @@ import helmet from "helmet"
 import userRoutes from "./app/api/user/router.js"
 import masterSizeRoutes from "./app/api/masterSize/router.js"
 import masterProductRoutes from "./app/api/masterProduct/router.js"
+import productPriceRoutes from "./app/api/productPrice/router.js"
 
 dotenv.config()
 
@@ -23,5 +24,6 @@ const URL = `/api/v1`
 app.use(`${URL}`, userRoutes);
 app.use(`${URL}`, masterSizeRoutes);
 app.use(`${URL}`, masterProductRoutes);
+app.use(`${URL}`, productPriceRoutes);
 
 app.listen(port, () => console.log(`running ⚡ port: ${port}`))